long => int

+++ csharp/gigasecond/Gigasecond.cs
@@ -4,7 +4,7 @@ public static class Gigasecond
 {
     public static DateTime Add(DateTime moment)
     {
-        const long gigasecond = 1_000_000_000;
+        const int gigasecond = 1_000_000_000;
         return moment.AddSeconds(gigasecond);
     }
 }
